# Vectorize the cmath matrix operations