Free lunch time

## What's changed?

Just to help understanding what's going on with new matrix multiplication.

It should be noted that the cmath matrix is made of multiple arrays where each of them represents each **column** of matrix. 
Imagine that we do multiply 4x4 matrix **A** and **B** to obtain the matrix **C**. If we do the matrix multiplication as we usually do like the following figure, we will have to pick up the row of **A** which cannot be vectorized:

![image](https://github.com/acts-project/algebra-plugins/assets/63090140/ce72df7e-496f-492a-b53c-0745e1a1001a)

So I came up with an idea to do the matrix multiplication column-wisely by multiplying a column of **A** to the element of **B**:

```c++
     for (size_type k = 0; k < M; ++k) {
        C[j][k] += A[i][k] * B[j][i];
      }
```

The following figure shows how the column of **C** is calculated in the new matrix multiplication

![image](https://github.com/acts-project/algebra-plugins/assets/63090140/c7346611-84f7-4751-997c-dbc89bcbe825)