@niermann999 2024-04-11
Is this single or double precision? You may not see much of a speedup in double precision