@stephenswat 2024-04-11
Are you compiling with support for vector extensions enabled?