@beomki-yeo 2024-04-11
Maybe not. I thought Release build would do that. Do you know how to enable it?