@niermann999 2024-04-11
You can also ask the compiler to tell you about vectorization optimizations it did ```-fopt-info-vec-all``` and similar