@beomki-yeo 2024-04-15
Intel advisor report of `algebra_test_array` (compiled with `"-march=native" "-ftree-vectorize"`) says that the matrix multiplication is vectorized: 

![image](https://github.com/acts-project/algebra-plugins/assets/63090140/c45c7ee3-f4a2-4a33-8c5f-3e9726ddf27b)

And as expected, it is reported that there are bunch of non-vectorized operations with `partial_pivot_lud` (Example: https://github.com/acts-project/algebra-plugins/blob/9e684de5661ca9c1c216d0631e24951ffa10d9c9/math/cmath/include/algebra/math/algorithms/matrix/inverse/partial_pivot_lud.hpp#L67)

![image](https://github.com/acts-project/algebra-plugins/assets/63090140/a234442f-1163-4f12-baec-1e8f777c815f)

Partial Pivot LU decomposition is written totally anti-vectorized manner: The sub vector matrix represents a column of the matrix and we do the Gaussian elimination row-wisely. We will need to rewirte the partial pivot lud later with column-wise gauss elimination.